Trim examples, adding license when downloading data

+++ pyglmnet/datasets.py
@@ -11,7 +11,24 @@ from urllib.request import urlretrieve
 from tempfile import TemporaryDirectory
 
 pbar = None
+_rgcs_license_text = """
+License
+-------
 
+This tutorial dataset (RGC spikes data) is granted by
+original authors (E.J. Chichilnisky and Jonathan Pillow).
+The dataset ramains a property of the original authors.
+Its use and transfer outside tutorial, e.g. for research purposes, 
+is prohibited without written consent to the original authors.
+
+If you reference this dataset in your publications, please:
+    1) acknowledge its authors: E.J. Chichilnisky and Jonathan Pillow
+    2) cite thier publications as indicated on the tutorial:
+
+If you want to use it beyond the educational
+purposes or have further questions, please contact
+Jonathan Pillow (pillow@princeton.edu).
+"""
 
 def _reporthook(count, block_size, total_size):
     """Report download percentage."""
@@ -101,7 +118,7 @@ def fetch_community_crime_data():
     return X, y
 
 
-def fetch_RGCs_data(dpath):
+def fetch_RGCs_data(dpath, accept_rgcs_license=False):
     """
     Downloads data for spike trains prediction in retinal ganglia cells.
     Please see https://github.com/glm-tools/datasets/RGCs/ for permission
@@ -117,12 +134,19 @@ def fetch_RGCs_data(dpath):
     dpath : str
         The data path
     """
+    if accept_rgcs_license:
+        answer = 'y'
+    else:
+        answer = input('%s\nAgree (y/[n])? ' % _rgcs_license_text)
+    if answer.lower() != 'y':
+        raise RuntimeError('You must agree to the license to use this '
+                            'dataset')
 
     if os.path.exists(dpath):
         shutil.rmtree(dpath)
     os.mkdir(dpath)
 
-    base_url = "https://raw.githubusercontent.com/glm-tools/datasets/master"
+    base_url = "https://raw.githubusercontent.com/titipata/datasets/rgcs_dataset"
     fnames = ['data_RGCs.json']
 
     for fname in fnames:
